merge V4.0.0 into branch

See merge request morvaknin/trivia_gali_yuval_2024!72

+++ Server/RoomData.h
@@ -14,6 +14,6 @@ struct RoomData
 	std::string owner;
 	unsigned int maxPlayers;
 	unsigned int numOfQuestionsInGame;
-	double timePerQuestion;
+	unsigned int timePerQuestion;
 	unsigned int isActive;
 };
